+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* Use Github release to publish vagrant files
+* Upgrade rest-interface version to 0.1.1
+* New demo: Add Vagrantfile for Ubuntu VM
+* Other minor updates